# docs/swagger.yaml
definitions:
  swagger.SwaggerAccessResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  swagger.SwaggerNewError:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /access/{guid}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Идентификатор пользователя
        in: path
        name: guid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SwaggerAccessResponse'
        "400":
          description: Ошибка создания токена
          schema:
            $ref: '#/definitions/swagger.SwaggerNewError'
        "500":
          description: Ошибка парсинга UserAgent
          schema:
            $ref: '#/definitions/swagger.SwaggerNewError'
      summary: Получить пару access, refresh токенов
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"